requirements.txt
molo.core==6.7.0
celery<4.0
molo.surveys==6.4.6
molo.yourwords==6.2.5
molo.commenting==6.2.5
psycopg2==2.7.1
django-modelcluster>=2.0,<3.0
html5lib==0.9999999
django_compressor==2.1
responses
Unidecode==0.04.16
django-storages
boto
elasticsearch>=2.0.0,<3.0.0
freezegun
